// gen/models/wifi-info.ts
// tslint:disable
/**
 * 实训打卡
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * WiFi校验信息
 * @export
 * @interface WifiInfo
 */
export interface WifiInfo {
    /**
     * WiFi名称
     * @type {string}
     * @memberof WifiInfo
     */
    ssid: string;
    /**
     * WiFi MAC地址
     * @type {string}
     * @memberof WifiInfo
     */
    bssid: string;
}