// gen/models/task-verification-config.ts
// tslint:disable
/**
 * 实训打卡
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { CheckinMethods } from './checkin-methods';
import { NfcInfo } from './nfc-info';
import { TaskVerificationConfigLocationInfo } from './task-verification-config-location-info';
import { WifiInfo } from './wifi-info';

/**
 * 任务校验配置组件，包含校验方式配置和相关参数
 * @export
 * @interface TaskVerificationConfig
 */
export interface TaskVerificationConfig {
    /**
     * 
     * @type {CheckinMethods}
     * @memberof TaskVerificationConfig
     */
    checkinMethods: CheckinMethods;
    /**
     * 
     * @type {TaskVerificationConfigLocationInfo}
     * @memberof TaskVerificationConfig
     */
    locationInfo: TaskVerificationConfigLocationInfo;
    /**
     * 
     * @type {WifiInfo}
     * @memberof TaskVerificationConfig
     */
    wifiInfo?: WifiInfo;
    /**
     * 
     * @type {NfcInfo}
     * @memberof TaskVerificationConfig
     */
    nfcInfo?: NfcInfo;
}