// gen/models/checkin-task.ts
// tslint:disable
/**
 * 实训打卡
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { TaskVerificationConfig } from './task-verification-config';

/**
 * 
 * @export
 * @interface CheckinTask
 */
export interface CheckinTask {
    /**
     * 签到任务ID
     * @type {number}
     * @memberof CheckinTask
     */
    taskId: number;
    /**
     * 所属用户组ID
     * @type {number}
     * @memberof CheckinTask
     */
    groupId: number;
    /**
     * 任务名称
     * @type {string}
     * @memberof CheckinTask
     */
    taskName: string;
    /**
     * 任务描述
     * @type {string}
     * @memberof CheckinTask
     */
    description?: string;
    /**
     * 签到开始时间（Unix时间戳，单位：秒）
     * @type {number}
     * @memberof CheckinTask
     */
    startTime: number;
    /**
     * 签到结束时间（Unix时间戳，单位：秒）
     * @type {number}
     * @memberof CheckinTask
     */
    endTime: number;
    /**
     * 任务状态
     * @type {string}
     * @memberof CheckinTask
     */
    status: CheckinTaskStatusEnum;
    /**
     * 
     * @type {TaskVerificationConfig}
     * @memberof CheckinTask
     */
    verificationConfig: TaskVerificationConfig;
    /**
     * 创建时间（Unix时间戳，单位：秒）
     * @type {number}
     * @memberof CheckinTask
     */
    createdAt?: number;
}

/**
    * @export
    * @enum {string}
    */
export enum CheckinTaskStatusEnum {
    Ongoing = 'ongoing',
    Upcoming = 'upcoming',
    Expired = 'expired'
}