// gen/models/verification-data.ts
// tslint:disable
/**
 * 实训打卡
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { NfcInfo } from './nfc-info';
import { VerificationDataLocationInfo } from './verification-data-location-info';
import { WifiInfo } from './wifi-info';

/**
 * 校验数据组件，根据不同的校验方式需要提供不同的字段
 * @export
 * @interface VerificationData
 */
export interface VerificationData {
    /**
     * 
     * @type {VerificationDataLocationInfo}
     * @memberof VerificationData
     */
    locationInfo?: VerificationDataLocationInfo;
    /**
     * 人脸识别数据的Base64编码（仅当任务需要人脸校验时必须提供）
     * @type {string}
     * @memberof VerificationData
     */
    faceData?: string;
    /**
     * 
     * @type {WifiInfo}
     * @memberof VerificationData
     */
    wifiInfo?: WifiInfo;
    /**
     * 
     * @type {NfcInfo}
     * @memberof VerificationData
     */
    nfcInfo?: NfcInfo;
}